test: update api token manage test
