--- tests/Feature/ApiTokenPermissionsTest.php
@@ -1,51 +1,0 @@
-<?php
-
-namespace Tests\Feature;
-
-use App\Models\PersonalAccessToken;
-use App\Models\User;
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Support\Str;
-use Laravel\Jetstream\Features;
-use Laravel\Jetstream\Http\Livewire\ApiTokenManager;
-use Livewire\Livewire;
-use Tests\TestCase;
-
-class ApiTokenPermissionsTest extends TestCase
-{
-    use RefreshDatabase;
-
-    public function test_api_token_permissions_can_be_updated(): void
-    {
-        if (! Features::hasApiFeatures()) {
-            $this->markTestSkipped('API support is not enabled.');
-        }
-
-        /** @var User */
-        $user = User::factory()->withPersonalTeam()->create();
-
-        $this->actingAs($user);
-
-        $token = $user->tokens()->create([
-            'name' => 'Test Token',
-            'token' => Str::random(40),
-            'abilities' => ['create', 'read'],
-        ]);
-
-        Livewire::test(ApiTokenManager::class)
-            ->set(['managingPermissionsFor' => $token])
-            ->set(['updateApiTokenForm' => [
-                'permissions' => [
-                    'delete',
-                    'missing-permission',
-                ],
-            ]])
-            ->call('updateApiToken');
-
-        /** @var PersonalAccessToken */
-        $token = $user->fresh()?->tokens->first();
-        $this->assertTrue($token->can('delete'));
-        $this->assertFalse($token->can('read'));
-        $this->assertFalse($token->can('missing-permission'));
-    }
-}
